IssueNo:#I57VNP:fix 64 fuzzer Description:fix 64 fuzzer Sig:SIG_ApplicaitonFramework Feature or Bugfix:Bugfix Binary Source:No

+++ test/fuzztest/getapplicationinfos_fuzzer/getapplicationinfos_fuzzer.cpp
@@ -56,7 +56,8 @@ namespace OHOS {
         }
 
         std::vector<ApplicationInfo> bundleInfos;
-        return bundleMgr->GetApplicationInfos(reinterpret_cast<int>(data), reinterpret_cast<int>(data), bundleInfos);
+        return bundleMgr->GetApplicationInfos(
+            reinterpret_cast<uintptr_t>(data), reinterpret_cast<uintptr_t>(data), bundleInfos);
     }
 }
